Improve the function of allowing users to change their profile pictures.

+++ src/main/java/com/melodiousplayer/controller/SysUserController.java
@@ -123,15 +123,18 @@ public class SysUserController {
     @PreAuthorize("hasAuthority('system:user:edit')")
     public R updateAvatar(@RequestBody SysUser sysUser) {
         SysUser currentUser = sysUserService.getById(sysUser.getId());
-        String avatarPath = avatarImagesFilePath + currentUser.getAvatar();
-        File avatarFile = new File(avatarPath);
-        if (avatarFile.exists()) {
-            boolean deleted = avatarFile.delete();
-            if (!deleted) {
-                return R.error("用户头像删除失败");
+        String defaultAvatar = "default.jpg";
+        if (!currentUser.getAvatar().equals(defaultAvatar)) {
+            String avatarPath = avatarImagesFilePath + currentUser.getAvatar();
+            File avatarFile = new File(avatarPath);
+            if (avatarFile.exists()) {
+                boolean deleted = avatarFile.delete();
+                if (!deleted) {
+                    return R.error("用户头像删除失败");
+                }
+            } else {
+                return R.error("用户头像不存在：" + currentUser.getAvatar());
             }
-        } else {
-            return R.error("用户头像不存在：" + currentUser.getAvatar());
         }
         currentUser.setUpdateTime(new Date());
         currentUser.setAvatar(sysUser.getAvatar());
